requirements.txt
termcolor
timber
colorama
docopt
decopts==1.0.2


aniso8601
click
cycler
Cython
Flask
Flask-SocketIO
gevent
greenlet
gunicorn
itsdangerous
Jinja2
MarkupSafe
matplotlib
numpy
olefile
opencv-python
Pillow
pyparsing
pytesseract
python-dateutil
python-engineio
python-socketio
pytz
selenium
six
SQLAlchemy
Werkzeug

